requirements.txt
python-dotenv
chromadb
sentence-transformers
tqdm
pyyaml
pymupdf
google-genai
requests
